play stand/walk animation based on entity state
we store the name of every animation now inside the `Animation`, so we
can checke which animation is currently running. this way, we can avoid
restarting an animation if it is already running inside
`play_animation`. the `Rc<String>` allows for cheaper clones of
`Animation`, and the name is immutable anyways.