add a mock udp socket that simulates packet loss, latency & jitter
this will make testing easier. also fixes some evil bugs in the unsafe
code, where the buffer-length was not reset in case of a panic, which
would then try to drop the invalid content in the buffer. (in our case,
since we only saved bytes, this didn't matter, but better be safe!)